Break up source for compression into more files.
